begin separating ContactUs page into sections
